+++ src/description/components/Color.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import SelectField from 'material-ui/SelectField';
 import MenuItem from 'material-ui/MenuItem';
 import { name as enumName, combined } from './../../enum';
-import { onSelectEnumChange } from './../../forms/onChange';
+import { onSelectEnumChange } from './../../forms/events';
 
 const Color = ({
   colors, name, onChange, values,
